% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{chatgpt}
\alias{chatgpt}
\title{Define a new class by injecting it into the global R environment.}
\usage{
chatgpt(name, members)
}
\description{
Define a new class by injecting it into the global R environment.
}
