% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{Class}
\alias{Class}
\title{Define a new RS class.}
\usage{
Class(.classname, ...)
}
\arguments{
\item{.classname}{The name of the class.}

\item{...}{The fields and methods of the class definition.}

\item{.validate}{Whether to validate the class attributes.
Note: setting this to \code{FALSE} gives a decent performance boost.}
}
\description{
Create a new ClassInstance in R.
}
\details{
The Class function creates a new ClassInstance in R.
It allows you to define fields and methods for the class.
}
