% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az.R
\name{:=}
\alias{:=}
\title{Declare a new class attribute.}
\usage{
`:=`(lhs, rhs)
}
\arguments{
\item{lhs}{The name of the attribute.}

\item{rhs}{The type of the attribute.}
}
\description{
Declare a new class attribute using the \verb{:=} operator.
}
