% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{\%:\%}
\alias{\%:\%}
\title{Type assignment operator}
\usage{
x \%:\% y
}
\arguments{
\item{x}{The variable to assign the type to.}

\item{y}{The type to assign to the variable.}
}
\value{
The variable with the assigned type.
}
\description{
\verb{\%:\%} is a type assignment operator that allows you to assign a type to a variable.
}
