% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorators.R
\name{private}
\alias{private}
\title{Declare a method or attribute as private.}
\usage{
private(.attr)
}
\arguments{
\item{.attr}{The function or attribute to be declared as private.}
}
\description{
Declare a method or attribute as private.
}
\details{
The \code{private} function is used to declare a method or attribute as private
in a class definition.
Private methods and attributes are not accessible from outside the class.
}
