% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{private_}
\alias{private_}
\title{Decorator to mark a class attribute as private.}
\usage{
private_(attribute)
}
\arguments{
\item{attribute}{The function or attribute to be declared as private.}
}
\description{
Declare a method or attribute as private.

The \code{private} function is used to declare a method or attribute as private
in a class definition.

Private methods and attributes are not accessible from outside the class,
and are used to encapsulate data that should not be modified directly.
}
