% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS-package.R
\docType{package}
\name{RS-package}
\alias{RS}
\alias{RS-package}
\title{RS: Fast classes for R}
\description{
Fast classes for R
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/avhz/RS/}
  \item \url{https://avhz.github.io/RS/}
}

}
\keyword{internal}
