% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorators.R
\name{static}
\alias{static}
\title{Declare a method is a static method.}
\usage{
static(.attr)
}
\arguments{
\item{.attr}{The function to be declared as a static method.}
}
\description{
Declare a method is a static method.
}
\details{
The \code{static} function is used to declare a method as a static method
in a class definition.
Static methods do not refer to \code{self},
i.e. the first argument is not the instance of the class.
}
