% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{static_}
\alias{static_}
\title{Decorator to mark a method as static.}
\usage{
static_(attribute)
}
\description{
A static method is a method that belongs to the class itself,
rather than to instances of the class.
It can be called without creating an instance of the class,
and it does not have access to instance-specific data (self).
}
