% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{structure_}
\alias{structure_}
\title{Function to structure an R object with a class and attributes.}
\usage{
structure_(robj, class, attribs)
}
\description{
Function to structure an R object with a class and attributes.
}
\keyword{internal}
