% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{Class}
\alias{Class}
\title{Class}
\usage{
Class(.classname, ...)
}
\arguments{
\item{...}{The fields and methods of the class.}

\item{name}{The name of the class.}
}
\description{
Create a new class
}
\details{
The Class function creates a new class in R.
It allows you to define fields and methods for the class.
}
