% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{private_}
\alias{private_}
\title{Decorator to mark a class attribute as private.}
\usage{
private_(attribute)
}
\description{
A private attribute is not accessible from outside the class.
It is used to encapsulate data that should not be modified directly.
}
